backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
groq==0.9.0
openai==1.30.1
gTTS==2.5.1
moviepy==1.0.3
Pillow==10.3.0
numpy==1.26.4
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-pptx==0.6.23
deep-translator==1.11.4
authlib==1.3.0
itsdangerous==2.2.0
